[test](fe) Isolate cloud frontend service test
### What problem does this PR solve?

Issue Number: None

Related PR: None

Problem Summary: FrontendServiceImplTest starts a non-cloud mocked FE and BE. testGetTabletReplicaInfosNullJobReturnsCancelledWithoutNpe temporarily enabled cloud mode through Config.cloud_unique_id, allowing the shared background heartbeat thread to take a cloud-only path and mark the mock BE dead. Move this cloud-specific test into a dedicated class without the mocked FE/BE fixture, while preserving the same getTabletReplicaInfos cancellation assertions.

### Release note

None

### Check List (For Author)

- Test: Unit Test
    - ./run-fe-ut.sh --run org.apache.doris.service.FrontendServiceImplCloudTest
    - ./run-fe-ut.sh --run org.apache.doris.service.FrontendServiceImplTest
- Behavior changed: No
- Does this need documentation: No
